requirements.txt
fastapi
uvicorn
numpy
torch
torchaudio
pydantic
omegaconf
pytest
python-multipart
ffmpeg-python
python-dotenv
websockets
soundfile
nemo_toolkit[asr]
livekit-agents
